refactor: 로깅 구현 방식 변경
